Fix petty hidden CSS rules
This one had to be complained about in its own commit. So check this.

I have buttons in a nav element that have flexed to fit their parent.
But for whatever reason, the nav element occupies more width than is
expected. Why? Because when I say in the reset that the button has a
width of 150px it retains that implicit definition, even though the
button itself may not have a width of 150px. So it gives me ghost width
values that I have to fix explicitly by setting `width: auto`. That is
annoying and borderline abusive and it needs to be fixed.