Correct some CSS
Made the heading elements bold.

Also figured out how to make the images white. NB: That inversion only
worked because the images were black, not because I found a way to layer
white onto them.

Lastly, made the buttons on the top row wrap onto a new line on thinner
screens. May be useless if/when I actually add media queries for 100%
responsiveness.