Create a work of art (#1)
* Create index.html file

* Create basic HTML layout

It's probably fine.

* Add SVG icons

* Create styles.css

* Rearrange icons to be contained in link

* Use header, main, footer format

* Change alt attributes to screen reader friendly

I think this is how this gets done. Probably. Anyway, I need to commit
some changes before something bad happens and I lose all my work.

* Add generic profile picture

This is for testing and should be changed once everything is in working
order.

* Insert basic styling for index.html

This is again, necessary so that I don't lose any work.

* Fix petty hidden CSS rules

This one had to be complained about in its own commit. So check this.

I have buttons in a nav element that have flexed to fit their parent.
But for whatever reason, the nav element occupies more width than is
expected. Why? Because when I say in the reset that the button has a
width of 150px it retains that implicit definition, even though the
button itself may not have a width of 150px. So it gives me ghost width
values that I have to fix explicitly by setting `width: auto`. That is
annoying and borderline abusive and it needs to be fixed.

* Add the spinning GitHub link

* Make a meagre attempt towards modular CSS

* Fix border radius and font sizes

* Change font to Roboto

* Use downloaded font

Something is wrong with the Google Fonts API. It could be the browser
or something, but I can't be bothered with it now.

* Add backup font

* Correct some CSS

Made the heading elements bold.

Also figured out how to make the images white. NB: That inversion only
worked because the images were black, not because I found a way to layer
white onto them.

Lastly, made the buttons on the top row wrap onto a new line on thinner
screens. May be useless if/when I actually add media queries for 100%
responsiveness.

* Change grid columns

Why? Just because.

Also made an honest attempt at responsiveness and it's not working. I'll
leave that for next time.

* Refactor CSS into variables

* Edit CSS variables

Oh jolly days! I actually found a shade of blue that works with the
thing. I also changed the background color of something somewhere so
that's noteworthy as well.

* Make the links crawlable

I changed the links to redirect to the home page whenever they are
clicked. I also removed the spaces I put earlier because VSCode can
shrink code blocks anyway.

* Add LightHouse recommendations

Made the headings consistent by using h4 and h5.

Added meta description and theme color. How's the theme color relevant?
Idk

* Add a version of responsiveness

I mean, it's all good now, right?

* Add basic shortcut icon

Note: is likely to change.

* Stop weird auto-zoom artefact on mobile.

I really don't get this. I don't. What's going on?

* Revert the flex display on the body

Something sketchy happened on an actual device. Fantastic 🙄

* Implement dark mode

Not my best work, but it is functional and keeps my 100 for 100 streak
so I'll take it.

* Fix renamed variable

* Change font size of h5 element

The user agent stylesheet strikes once more. Fantastic🤦

* Change file structure for profiles

* Add images for uniqueness

* Add some personality

It's hard to overstate how weird the tone of this dashboard is. Just
change the live branch to basic if I have to get a job somewhere. Or
just do a better job some other day.

* Update README.md

My descriptions of things are so silly sometimes. But they're lovely
nonetheless